Make drag and drop work.
It works but only intermittently. The idea
is supposed to be that you can drag a file from
goon to Finder/Explorer or other apps and it will
do the thing.

Not sure what needs to happen here to make it not
intermittant. Like the first drag doesn't work.
Maybe user-select needs to be none?