Fix rotated zoom issue.
The way it works now. First the image is rotated, then
the stretch mode is applied. This means it will stretch
as though it was rotated in the image file itself.

In other words, if you're in "fit to width" mode then, if you
rotate 90 degrees the "new width" (the rotated width) will be
used to fit.

```
original    rotated
with        with
width       width
used        used
marked      marked

+---+
|   |      +---------+
|<->|      |<------->|
|   |      +---------+
+---+

```